requirements.txt
Flask
pdf2pptx
docxcompose==1.0.0a1
pdfkit
PyMuPDF>=1.19.0
python-docx>=0.8.10
fonttools>=4.24.0
numpy>=1.17.2
opencv-python>=4.5
fire>=0.3.0